Fix to PageContent build issues
* Renamed the exported Props

HCRC-13. Because of some build errors, the props are renamed to more be more unique and so more followable.

* Export named functions only from PageContent

HCRC-13. Page content had some issues in a rollup build process. When the HCRC lib was used, the props from Text-component were used instead of the PageContentProps. Exporting the PageContent as a named function instead of default fixed the issue.