fix: convert booleans presented in string format to actual booleans
HH-254.

It's clear that there has been a bug in the normalized params usage, 
because the eventType was already handled, but the mutated params-object
was actually never used.

The boolean strings "true / "false" are now handled in that same params-object.
Also the transforming function is moved to the file meant for utils related 
to the Collection-component.