Support iframes for video media in HtmlToReact component
KK-1001.

Add a new as-secure-as-possible iframe component 
for CMS' embedded video content. The Wordpress headless CMS has 
content blocks for Youtube and Vimeo video medias that 
we should support. The HtmlToReact-component now has a new attribute that can be used to list the trusted (source) origins. The src-attribute of the iframe-elements should  include one of the given trusted origins.

Refactor the component replace functions in new files under a new folder.

Add some new utils for HtmlToReact. Serve them from a new utils folder,
so that they don't grow the `HtmlToReact.tsx` file. They are not added 
in the "global" utils folder, because they are so tide to the `HtmlToReact`.