fix: change lodash to lodash-es in order to fix the bundle issues
HH-234

The HCRC-lib was build properly and the storybook could be used,
but when it was used as a plugin in an app,
the app whined about that there were commonjs `module.exports`
inside the bundled esm javascript.
The issues was in the lodash-library that was used as a dependency and
was added quit recently. The lodash is now replaced with lodash-es,
which should fit better in esm and nextjs world.