fix: image link should search img-tag deeper from the children
HCRC-112.

Add a recursively map utility, that can be used to recursively map the React children.

Add a findAllElementsOfType utility, that takes advantage of the `recursiveMap` and `getChildrenByType` utilities.

Add tests to each utility.

Finally, the Link should use the new findAllElementsOfType utility to search for images inside an anchor.