fix: image sizes should be consistent and include size "large"
LIIKUNTA-700.

The featured image document has a field `featuredImage.mediaItemUrl`
which stores a value to the source of the original image uploaded to the
CMS media library. Nowadays the media library offers also some
combressed versions of the image: large, medium_large, medium and
thumbnail. The original media might be huge measured in a size of the
file, so it should not be used in web. However, the size "large" is
nowhere in use and the medium_large that is used, is usually too weak in
quality, to be used for example as a hero image.

This commit adds the large size and makes the featuredIamge document
consistent, no matter where it used, it will now offer all the file
sizes. It might be bad practise to request sizes of image that are not
used, but this library is generic and it makes the development much much
easier. Also the added slowness should not be "much".