fix(types): use strict type checking and fix all typing issues
HCRC-178.

Use strict type rules. 

Enable skipLibCheck.

Render null when Head-component is not configured. Fix the nullpointer
exceptions.

The LoadingButton is a basic HDS-React Button, which analyses isLoading
state to show a loading spinner as a start icon and changes styles
according to state as well. It mimics the example given in the HDS-React Storybook documentation in
https://hds.hel.fi/components/button/code/#loading-button.

The `handleEscapeKeyPress` cannot be a native KeyboardEvent and
React.KeyboardEvent at the same time. If a `onKeyDown`-attribute is
used, it needs to use the React-KeyboardEvent, but it should be exactly
the same as adding an eventListener to KeyboardEvent.

Added `classnames`-package to dependencies as it was needed by
Link-component.
test: update snapshots

HCRC-178

fix(types): all stories default export argTypes typing

HCRC-178

fix(types): simplify and fix useEventsByIdsQuery type handling

Remove makeQueryWithEventsApolloClientFromConfig.ts as unnecessary,
it was used in a single place. Now the code is simpler and easier to
understand.

Using HelmetWrapper the @ts-ignore clauses become unnecessary and are
removed.