refactor(carousel): add carousel context and split carousel components
HH-380.

Split the long and unclear carousel code to multiple subcomponents.

Create a new carousel context to serve all carousel related properties
from a context provider with a use of single context hook. By using a
context provider, it's easier to provide properties to subcomponents.
Without prop drilling, it's easier to understand the dependencies.

This commit is backward compatible, since the carousel component is
exported and used the same way as was earlier.