fix(carousel): add region roles and aria labels to carousel
HH-380.

To fix or improve carousel component accessibility, the carousel should
have a role region with a aria-label, that describes the wholeness of
the component: what is the title of the carousel and how many items does
it have.

Add region role and aria label to carousel slider. When navigating to
the carousel slider component, the screen reader should read the current
slide of the slider, the total amount of items and how many of the items
are currently shown.