feat: switch page content order when using small screen
LIIKUNTA-759.

Take advantage of grid styles to reorder content when a small screen
breakpoint is window size reached.

When using bigger than mobile sized screen, the content order would be:
1. page main (text) content
2. share buttons,
3. sidebar
4. modules, footer, etc...

When using mobile sized screen, switch the sidebar with secondary data
(share buttons...) in main content, so in mobile the
order would be:
1. page main (text) content
2. sidebar
3. share buttons,
5. modules, footer, etc...