ci: release-please with common workflows for canary and stable release
HCRC-189.

1. Use common workflow to release canary on mainbranch push.
2. Merge npm-publish-canary workflow to release-please.

The common Github workflows are using NPM's trusted publisher feature. A
Github workflow / action can be set as a trusted publisher, but the NPM
supports only 1 workflow. We have had 2: npm-canary-release and
npm-stable-release. After refactoring, the stable releases are done with
a release-please action, but we need to support releasing
canary-releases in the same workflow file, in order to publish both
types as a trusted publisher.