fix: disentangle customRender & testing-library imports from each other
Fix linting errors:
````
error  Multiple exports of name 'render'  import/export
```
and refactor imports/exports related to customRender function,
@testing-library/react and @testing-library/user-event. Previously
these were entangled, now they're clearer related to what is imported
from where rather.

NOTE: The renaming of customRender function to render function on import
is intentionally left as an explicit act so it is clear that it is being
done.

refs HCRC-156